Change `usize` field of `IntrusiveAlias` to `* const ()`.
Rust is moving towards a pointer provenance model that strongly
discourages casting pointer-sized integers to or from pointers.
To adapt to this, we change the field of the `IntrusiveAlias` type
from `usize` to `* const ()`. In making this change, I noticed
that the old code had an incorrect usage of Rust's memory safety
model: in the previous version, I considered _taking_ a raw pointer
(or our version of one) to be an unsafe operation. Rust, however,
does not consider this unsafe. It's _dereferencing_ a raw pointer
that's considered unsafe. I've aligned this code with Rust's
definition of `unsafe`, and in general, updated to more modern
Rust conventions.